fix(item) save관련 repository setting 및 controller 수정
Former-commit-id: 9f0d000d74daf06912357863f3ae1b2c6d7453b3

+++ backend/bangbang/src/main/java/com/bangbang/domain/item/ItemRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemRepository extends JpaRepository<Item, Integer> {
+public interface ItemRepository extends JpaRepository<Item, Long> {
     List<Item> findTop100By();
-    Item findById(int itemId);
+    Item findById(long itemId);
 }
